moved pg to gymlib
